Agentic Project/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
requests==2.32.3
Pillow==11.0.0
moviepy==2.1.1
pytest==8.3.3
